修复: PR #410 代码审查 — 消除 isSecureRequest 和 cookie 解析重复
- 提取 isSecureRequest() 到 src/utils.ts,消除 middleware/auth.ts 和 routes/auth.ts 中的重复实现
- 导出 getAllCookieValues() 和 tryVerifyAny() 从 middleware/auth.ts,并在 web.ts 的 WebSocket upgrade handler 中复用,去除内联重复逻辑
- WebSocket 路径不支持 Set-Cookie,忽略 needsUpgrade 标记(下次 HTTP 请求会完成升级)
