+++ src/web.ts
@@ -34,7 +34,11 @@ import {
 } from './schemas.js';
 
 // Middleware
-import { authMiddleware } from './middleware/auth.js';
+import {
+  authMiddleware,
+  getAllCookieValues,
+  tryVerifyAny,
+} from './middleware/auth.js';
 
 // Route modules
 import authRoutes from './routes/auth.js';
@@ -73,7 +77,7 @@ import {
   getUserById,
   updateAgentContextInfo,
 } from './db.js';
-import { isSessionExpired, verifySessionToken } from './auth.js';
+import { isSessionExpired } from './auth.js';
 import type {
   NewMessage,
   WsMessageOut,
@@ -540,41 +544,27 @@ function setupWebSocket(server: any): WebSocketServer {
     }
 
     // Verify session cookie (HMAC signature + DB lookup)
-    // Extract all values for the session cookie name to handle duplicate cookies
-    // (browsers may keep both old unsigned and new signed cookies)
-    const cookieHeader = request.headers.cookie || '';
-    let allCookieValues: string[] = [];
-    const securePrefix = SESSION_COOKIE_NAME_SECURE + '=';
-    const plainPrefix = SESSION_COOKIE_NAME_PLAIN + '=';
-    for (const part of cookieHeader.split(';')) {
-      const trimmed = part.trim();
-      if (trimmed.startsWith(securePrefix)) {
-        allCookieValues.push(trimmed.slice(securePrefix.length));
-      }
-    }
+    // Reuse shared cookie parsing helpers to handle duplicate cookies
+    // (browsers may keep both old unsigned and new signed cookies).
+    // Note: WebSocket upgrade cannot return Set-Cookie, so we ignore
+    // `needsUpgrade` here — the next HTTP request will upgrade the cookie.
+    const cookieHeader = request.headers.cookie as string | undefined;
+    let allCookieValues = getAllCookieValues(cookieHeader, SESSION_COOKIE_NAME_SECURE);
     if (allCookieValues.length === 0) {
-      for (const part of cookieHeader.split(';')) {
-        const trimmed = part.trim();
-        if (trimmed.startsWith(plainPrefix)) {
-          allCookieValues.push(trimmed.slice(plainPrefix.length));
-        }
-      }
+      allCookieValues = getAllCookieValues(cookieHeader, SESSION_COOKIE_NAME_PLAIN);
     }
     if (allCookieValues.length === 0) {
       socket.write('HTTP/1.1 401 Unauthorized\r\n\r\n');
       socket.destroy();
       return;
     }
-    let token: string | null = null;
-    for (const cv of allCookieValues) {
-      token = verifySessionToken(cv);
-      if (token) break;
-    }
-    if (!token) {
+    const verifyResult = tryVerifyAny(allCookieValues);
+    if (!verifyResult) {
       socket.write('HTTP/1.1 401 Unauthorized\r\n\r\n');
       socket.destroy();
       return;
     }
+    const token = verifyResult.token;
 
     const session = getCachedSessionWithUser(token);
     if (!session) {
